requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
MetaTrader5==5.0.44
mongoengine==0.27.0
mysqlclient==2.1.1
pymongo==3.12.3
pytz==2022.7.1
redis==4.5.4
